Move Jinja2 function url_for_slug and url_for_slug_path in jssg/templatetags/
